requirements.txt
-i https://pypi.org/simple
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
flask==1.1.2
future==0.18.2
graphviz==0.14.1
idna==2.10
itsdangerous==1.1.0
jinja2==2.11.2
line-bot-sdk==1.17.0
markupsafe==1.1.1
pygraphviz==1.6
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
transitions==0.8.3
urllib3==1.25.10
werkzeug==1.0.1
